Moved the file. Added the mobile table view. Adjusted the position of the navigation buttons on mobile.
